+++ src/types/common.ts
@@ -1,4 +1,4 @@
-export interface response<T = unknown> {
+export type ApiResponse<T = unknown> = {
   code: number
   data: T
   msg: string