+++ src/types/tasks.ts
@@ -1,4 +1,0 @@
-export type Task = {
-  task_id: number
-  task_name: string
-}