+++ src/vite-env.d.ts
@@ -1,8 +1,5 @@
 /// <reference types="vite/client" />
 
-declare module '*.css';
-declare module '*.scss';
-
 interface ImportMetaEnv {
   readonly VITE_API_BASE_URL: string
 }